Extract component edit profile modal and edit avatar modal and adding some compound components with base edit profile modal wrapper component
