+++ src/components/edit-profile-modal-wrapper/EditProfileModalWrapper.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./EditProfileModalWrapper.scss";
+
+export default function EditProfileModalWrapper({ children }) {
+  return (
+    <div className="content-container edit-profile-and-avatar-modal">
+      <div className="modal-content-wrapper">{children}</div>
+    </div>
+  );
+}
