+++ src/components/edit-profile-modal-wrapper/EditProfileModalWrapper.scss
@@ -1,0 +1,36 @@
+.edit-profile-and-avatar-modal {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  padding: 60px 10px;
+  top: 0;
+  overflow: auto;
+  background-color: rgb(0, 0, 0); /* Fallback color */
+  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
+  z-index: 5;
+  direction: ltr;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  @media only screen and (min-width: 1024px) {
+    padding: 20px;
+  }
+
+  .modal-content-wrapper {
+    width: 100%;
+    padding: 10px;
+    background-color: #fefefe;
+    border-radius: 5px;
+    display: flex;
+    flex-direction: column;
+    margin: 10% auto;
+    font-size: 12px !important;
+
+    @media only screen and (min-width: 1024px) {
+      width: 500px;
+      padding: 20px;
+    }
+  }
+}
