+++ src/components/edit-profile-modal-wrapper/bottom-content/BottomContent.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import "./BottomContent.scss";
+
+export default function BottomContent({ children }) {
+  return <div className="bottom-content">{children}</div>;
+}
