+++ src/components/edit-profile-modal-wrapper/bottom-content/BottomContent.scss
@@ -1,0 +1,13 @@
+.bottom-content {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  gap: 10px;
+  align-items: center;
+  width: 100%;
+
+  @media only screen and (min-width: 1024px) {
+    gap: 0px;
+    justify-content: space-evenly;
+  }
+}
