+++ src/components/edit-profile-modal-wrapper/middle-content/MiddleContent.scss
@@ -1,0 +1,16 @@
+.middle-content {
+  width: 100%;
+  height: fit-content;
+
+  .input-profile-box {
+    border-radius: 10px;
+    display: flex;
+    flex-direction: column;
+    padding: 20px 0px;
+    font-weight: 500;
+
+    input {
+      font-size: 12px;
+    }
+  }
+}
