+++ src/components/edit-profile-modal/EditProfileModal.scss
@@ -1,208 +1,26 @@
-.edit-profile-modal {
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  padding: 60px 10px;
-  top: 0;
-  overflow: auto;
-  background-color: rgb(0, 0, 0); /* Fallback color */
-  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
-  z-index: 5;
-  direction: ltr;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
+.bottom-edit-button {
+  font-size: 10px;
+  font-weight: 500;
+  border-radius: 8px;
+  color: white;
+  cursor: pointer;
+  text-align: center;
+  padding: 10px 0px;
+  width: 200px;
+  border: none;
 
   @media only screen and (min-width: 1024px) {
-    padding: 20px;
+    font-size: 12px;
   }
 
-  .modal-content-wrapper {
-    width: 100%;
-    padding: 10px;
-    background-color: #fefefe;
-    border-radius: 5px;
-    display: flex;
-    flex-direction: column;
-    margin: 10% auto;
-    font-size: 12px !important;
+  &.cancel {
+    background-color: white;
+    color: rgb(231, 101, 101);
+    border: solid 1px rgb(231, 101, 101);
+  }
 
-    @media only screen and (min-width: 1024px) {
-      width: 500px;
-      padding: 20px;
-    }
-
-    .top-content {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-
-      .edit-profile-modal-content {
-        font-size: 14px;
-        font-weight: 500;
-      }
-      .modal-close-button {
-        cursor: pointer;
-      }
-    }
-
-    .middle-content {
-      width: 100%;
-      height: fit-content;
-
-      .input-profile-box {
-        border-radius: 10px;
-        display: flex;
-        flex-direction: column;
-        padding: 20px 0px;
-        font-weight: 500;
-
-        input {
-          font-size: 12px;
-        }
-
-        // .edit-profile-input {
-        //   height: 36px;
-        //   border-radius: 10px;
-        //   border: solid 1px gray;
-        //   font-size: 18px;
-        //   padding-left: 20px;
-        //   margin-bottom: 10px;
-
-        //   &:focus {
-        //     outline: none;
-        //   }
-        // }
-
-        .progress-upload-wrapper {
-          display: flex;
-          flex-direction: column;
-
-          .upload-progress-persen {
-            font-weight: 10px;
-            margin-bottom: 6px;
-          }
-
-          .progress-bar {
-            background-color: #42b72a;
-            border-radius: 5px;
-            height: 8px;
-          }
-        }
-
-        .edit-profile-url {
-          background-color: aqua;
-          padding: 10px 0px 10px 10px;
-          border-radius: 10px;
-          cursor: pointer;
-        }
-
-        .preview-img-container {
-          position: relative;
-          margin-bottom: 10px;
-
-          .preview-img {
-            width: 100%;
-            height: 270px;
-            object-fit: contain;
-          }
-
-          .cancel-preview-img {
-            position: absolute;
-            top: -10px;
-            right: 4px;
-            cursor: pointer;
-          }
-        }
-
-        // .inputErrorMessage {
-        //   margin-bottom: 10px;
-        //   margin-left: 20px;
-        //   color: red;
-        //   font-size: small;
-        // }
-
-        .register-button {
-          height: 50px;
-          border-radius: 10px;
-          border-style: none;
-          background-color: #2c2891;
-          color: white;
-          font-size: 20px;
-          font-weight: 500;
-          margin-bottom: 10px;
-          cursor: pointer;
-        }
-
-        .register-forgot {
-          text-align: center;
-          color: indigo;
-          cursor: pointer;
-          font-weight: 300;
-
-          &:hover {
-            font-weight: 500;
-          }
-        }
-
-        .register-login-button {
-          width: 70%;
-          align-self: center;
-          text-align: center;
-          border-radius: 10px;
-          border-style: none;
-          background-color: #42b72a;
-          color: white;
-          font-size: 20px;
-          font-weight: 500;
-          cursor: pointer;
-          padding: 8px 0px;
-          text-decoration: none;
-        }
-      }
-    }
-
-    .bottom-content {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      gap: 10px;
-      align-items: center;
-      width: 100%;
-      
-      @media only screen and (min-width: 1024px) {
-        gap: 0px;
-        justify-content: space-evenly;
-      }
-
-      .bottom-edit-button {
-        font-size: 10px;
-        font-weight: 500;
-        border-radius: 8px;
-        color: white;
-        cursor: pointer;
-        text-align: center;
-        padding: 10px 0px;
-        width: 200px;
-        border: none;
-
-        @media only screen and (min-width: 1024px) {
-          font-size: 12px;
-        }
-
-        &.cancel {
-          background-color: white;
-          color: rgb(231, 101, 101);
-          border: solid 1px rgb(231, 101, 101);
-        }
-
-        &.save {
-          background-color: #605beb;
-          border: solid 1px #605beb;
-        }
-      }
-    }
+  &.save {
+    background-color: #605beb;
+    border: solid 1px #605beb;
   }
 }
