+++ src/components/input-text-global-v2/InputTextGlobalV2.scss
@@ -129,6 +129,7 @@
 
       &.if-edit-profile-style-date {
         height: 36px;
+        font-size: 12px;
       }
 
       &.if-type-date {
